Fix nos papers mesage, move all content of papers to manuscripts card, change some layout
